refactor: Moved default color ramps to ColorRamp.ts
