+++ src/colorizer/index.ts
@@ -2,10 +2,10 @@ export type { IFrameLoader, IArrayLoader, ArraySource } from "./loaders/ILoader"
 import ImageFrameLoader from "./loaders/ImageFrameLoader";
 import JsonArrayLoader from "./loaders/JsonArrayLoader";
 
-import ColorRamp from "./ColorRamp";
+import { colorRamps, ColorRamp } from "./ColorRamp";
 import ColorizeCanvas from "./ColorizeCanvas";
 import Dataset from "./Dataset";
 import Track from "./Track";
 import Plotting from "./Plotting";
 
-export { ColorizeCanvas, Plotting, Dataset, Track, ColorRamp, ImageFrameLoader, JsonArrayLoader };
+export { ColorizeCanvas, Plotting, Dataset, Track, colorRamps, ColorRamp, ImageFrameLoader, JsonArrayLoader };
